fix(Nibiru-Home/sprints#55): Arreglo tests Payment

+++ src/main/java/tienda_back/domain/service/impl/CategoryServiceImpl.java
@@ -43,6 +43,9 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteById(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new ResourceNotFoundException("La categoria con el id: " + id + " no existe");
+        }
         categoryRepository.deleteById(id);
     }
